worker/pyproject.toml
[project]
name = "worker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "dEAdbEEF314", email = "deadbeef314@gmail.com" }
]
requires-python = ">=3.12"
dependencies = [
    "boto3>=1.42.89",
    "musicbrainzngs>=0.7.1",
    "mutagen>=1.47.0",
    "openai>=2.31.0",
    "pillow>=12.2.0",
    "pyacoustid>=1.3.1",
    "pydantic-settings>=2.13.1",
    "python-dotenv>=1.2.2",
    "requests>=2.33.1",
]

[build-system]
requires = ["uv_build>=0.10.12,<0.11.0"]
build-backend = "uv_build"
